Add hud and bullet collision
Adicionei uma funcao que desenha a HUD do jogador e adiciona a colisao com entidades, além de atualizar os status do player

+++ src/player.h
@@ -7,7 +7,7 @@
 
 typedef struct player{
     Vector2 position;
-    int fuel;
+    float fuel;
     int life;
     int score;
 } PLAYER;
